+++ src/features/author/selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from '../../app/store';
+
+export const authorSelector = (state: RootState) => state.author;
