+++ src/features/comments/selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from '../../app/store';
+
+export const commentsSelector = (state: RootState) => state.comments;
