+++ src/features/posts/selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from '../../app/store';
+
+export const postsSelector = (state: RootState) => state.posts;
