+++ src/features/selectedPost/selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from '../../app/store';
+
+export const selectedPostSelector = (state: RootState) => state.selectedPost;
