+++ src/features/users/selectors.ts
@@ -1,0 +1,3 @@
+import { RootState } from '../../app/store';
+
+export const usersSelector = (state: RootState) => state.users;
